+++ docs/src/changelog.md
@@ -18,6 +18,11 @@ If your currently installed *Quality-time* version is v4.10.0 or older, please r
 
 - Sorting measurement entities by status, status end date, and status rationale did not work. Fixes [#8508](https://github.com/ICTU/quality-time/issues/8508).
 
+### Added
+
+- When a report has a configured issue tracker, show a card in the dashboard with the number of issues per issue status (open, in progress, done). The card can be hidden via the Settings panel. Clicking the card or setting "Visible metrics" to "Metrics with issues" in the Settings panel hides metrics without issues. Closes [#8222](https://github.com/ICTU/quality-time/issues/8222).
+- Allow for hiding the legend card via the Settings panel.
+
 ## v5.11.0 - 2024-04-22
 
 ### Deployment notes